+++ worker/src/RTC/DirectTransport.cpp
@@ -146,7 +146,7 @@ namespace RTC
 	}
 
 	void DirectTransport::SendRtpPacket(
-	  RTC::Consumer* consumer, RTC::RTP::Packet* packet, const RTC::Transport::onSendCallback* cb)
+	  RTC::Consumer* consumer, RTC::RTP::Packet* packet, onSendCallback cb)
 	{
 		MS_TRACE();
 
@@ -156,8 +156,7 @@ namespace RTC
 
 			if (cb)
 			{
-				(*cb)(false);
-				delete cb;
+				cb(false);
 			}
 
 			return;
@@ -177,8 +176,7 @@ namespace RTC
 
 		if (cb)
 		{
-			(*cb)(true);
-			delete cb;
+			cb(true);
 		}
 
 		// Increase send transmission.
@@ -226,7 +224,7 @@ namespace RTC
 	}
 
 	void DirectTransport::SendMessage(
-	  RTC::DataConsumer* dataConsumer, RTC::SCTP::Message message, onQueuedCallback* cb)
+	  RTC::DataConsumer* dataConsumer, RTC::SCTP::Message message, onMessageQueuedCallback cb)
 	{
 		MS_TRACE();
 
@@ -245,8 +243,7 @@ namespace RTC
 
 		if (cb)
 		{
-			(*cb)(true, false);
-			delete cb;
+			cb(true, /*isSendBufferFull*/ false);
 		}
 
 		// Increase send transmission.