+++ worker/src/RTC/PipeTransport.cpp
@@ -505,7 +505,7 @@ namespace RTC
 	}
 
 	void PipeTransport::SendRtpPacket(
-	  RTC::Consumer* /*consumer*/, RTC::RTP::Packet* packet, RTC::Transport::onSendCallback* cb)
+	  RTC::Consumer* /*consumer*/, RTC::RTP::Packet* packet, onSendCallback cb)
 	{
 		MS_TRACE();
 
@@ -513,8 +513,7 @@ namespace RTC
 		{
 			if (cb)
 			{
-				(*cb)(false);
-				delete cb;
+				cb(false);
 			}
 
 			return;
@@ -527,14 +526,13 @@ namespace RTC
 		{
 			if (cb)
 			{
-				(*cb)(false);
-				delete cb;
+				cb(false);
 			}
 
 			return;
 		}
 
-		this->tuple->Send(data, len, cb);
+		this->tuple->Send(data, len, std::move(cb));
 
 		// Increase send transmission.
 		RTC::Transport::DataSent(len);
@@ -589,11 +587,11 @@ namespace RTC
 	}
 
 	void PipeTransport::SendMessage(
-	  RTC::DataConsumer* dataConsumer, RTC::SCTP::Message message, onQueuedCallback* cb)
+	  RTC::DataConsumer* dataConsumer, RTC::SCTP::Message message, onMessageQueuedCallback cb)
 	{
 		MS_TRACE();
 
-		SendSctpMessage(dataConsumer, std::move(message), cb);
+		SendSctpMessage(dataConsumer, std::move(message), std::move(cb));
 	}
 
 	bool PipeTransport::SendData(const uint8_t* data, size_t len)