+++ worker/src/RTC/PlainTransport.cpp
@@ -819,7 +819,7 @@ namespace RTC
 	}
 
 	void PlainTransport::SendRtpPacket(
-	  RTC::Consumer* /*consumer*/, RTC::RTP::Packet* packet, const RTC::Transport::onSendCallback* cb)
+	  RTC::Consumer* /*consumer*/, RTC::RTP::Packet* packet, onSendCallback cb)
 	{
 		MS_TRACE();
 
@@ -827,8 +827,7 @@ namespace RTC
 		{
 			if (cb)
 			{
-				(*cb)(false);
-				delete cb;
+				cb(false);
 			}
 
 			return;
@@ -841,14 +840,13 @@ namespace RTC
 		{
 			if (cb)
 			{
-				(*cb)(false);
-				delete cb;
+				cb(false);
 			}
 
 			return;
 		}
 
-		this->tuple->Send(data, len, cb);
+		this->tuple->Send(data, len, std::move(cb));
 
 		// Increase send transmission.
 		RTC::Transport::DataSent(len);
@@ -917,11 +915,11 @@ namespace RTC
 	}
 
 	void PlainTransport::SendMessage(
-	  RTC::DataConsumer* dataConsumer, RTC::SCTP::Message message, onQueuedCallback* cb)
+	  RTC::DataConsumer* dataConsumer, RTC::SCTP::Message message, onMessageQueuedCallback cb)
 	{
 		MS_TRACE();
 
-		SendSctpMessage(dataConsumer, std::move(message), cb);
+		SendSctpMessage(dataConsumer, std::move(message), std::move(cb));
 	}
 
 	bool PlainTransport::SendData(const uint8_t* data, size_t len)