+++ worker/src/RTC/TcpConnection.cpp
@@ -168,7 +168,7 @@ namespace RTC
 		}
 	}
 
-	void TcpConnection::Send(const uint8_t* data, size_t len, ::TcpConnectionHandle::onSendCallback* cb)
+	void TcpConnection::Send(const uint8_t* data, size_t len, onSendCallback cb)
 	{
 		MS_TRACE();
 
@@ -177,6 +177,6 @@ namespace RTC
 		uint8_t frameLen[2];
 
 		Utils::Byte::Set2Bytes(frameLen, 0, len);
-		::TcpConnectionHandle::Write(frameLen, 2, data, len, cb);
+		::TcpConnectionHandle::Write(frameLen, 2, data, len, std::move(cb));
 	}
 } // namespace RTC