+++ rust/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Worker: Fix endless regeneration of FlatBuffers generated headers ([PR #1926](https://github.com/versatica/mediasoup/pull/1926)).
 - SCTP: Fix unbounded SCTP reassembly queue growth during deferred reset processing ([PR #1927](https://github.com/versatica/mediasoup/pull/1927)).
+- Worker: Refactor send callbacks ([PR #1930](https://github.com/versatica/mediasoup/pull/1930)).
 
 ### 0.28.0